move get_random_array -> tests/common.py
